# Fix importer performance regression