**Description**:
<!--
One or two line summary of what this PR does and why it is needed, followed by a list
of changes in imperative, present tense for use in the commit message or changelog. Example:

This PR modifies ... in order to support ...
* Add config property
* Change column name
* Remove ...
-->

This PR fixes the importer performance regression caused by the new btree index over transaction_hash.

- Use postgresql hash index on transaction_hash

**Related issue(s)**:

Fixes #4308

**Notes for reviewer**:
<!-- Provide logs, performance numbers or screenshots of the new functionality -->

Tested the fix with perfnet record files:

```
2022-08-18T08:39:35.324-0600 INFO scheduling-3 c.h.m.i.p.b.BatchInserter Copied 20000 rows to transaction table in 260.0 ms
2022-08-18T08:39:36.252-0600 INFO scheduling-3 c.h.m.i.p.b.BatchInserter Copied 20000 rows to transaction table in 260.3 ms
2022-08-18T08:39:37.095-0600 INFO scheduling-3 c.h.m.i.p.b.BatchInserter Copied 20000 rows to transaction table in 262.8 ms
2022-08-18T08:39:37.730-0600 INFO scheduling-3 c.h.m.i.p.b.BatchInserter Copied 20000 rows to transaction table in 265.1 ms
2022-08-18T08:39:38.421-0600 INFO scheduling-3 c.h.m.i.p.b.BatchInserter Copied 20000 rows to transaction table in 262.9 ms
2022-08-18T08:39:39.464-0600 INFO scheduling-3 c.h.m.i.p.b.BatchInserter Copied 20000 rows to transaction table in 285.7 ms
2022-08-18T08:39:41.813-0600 INFO scheduling-3 c.h.m.i.p.b.BatchInserter Copied 20000 rows to transaction table in 260.1 ms
2022-08-18T08:39:42.504-0600 INFO scheduling-3 c.h.m.i.p.b.BatchInserter Copied 20000 rows to transaction table in 290.2 ms
2022-08-18T08:39:43.833-0600 INFO scheduling-3 c.h.m.i.p.b.BatchInserter Copied 20000 rows to transaction table in 260.8 ms
2022-08-18T08:39:44.830-0600 INFO scheduling-3 c.h.m.i.p.b.BatchInserter Copied 20000 rows to transaction table in 262.9 ms
2022-08-18T08:39:45.818-0600 INFO scheduling-3 c.h.m.i.p.b.BatchInserter Copied 20000 rows to transaction table in 264.6 ms
2022-08-18T08:39:46.684-0600 INFO scheduling-3 c.h.m.i.p.b.BatchInserter Copied 20000 rows to transaction table in 279.3 ms
2022-08-18T08:39:47.787-0600 INFO scheduling-3 c.h.m.i.p.b.BatchInserter Copied 20000 rows to transaction table in 403.4 ms
2022-08-18T08:39:48.714-0600 INFO scheduling-3 c.h.m.i.p.b.BatchInserter Copied 20000 rows to transaction table in 330.1 ms
2022-08-18T08:39:49.456-0600 INFO scheduling-3 c.h.m.i.p.b.BatchInserter Copied 20000 rows to transaction table in 269.3 ms
2022-08-18T08:39:50.280-0600 INFO scheduling-3 c.h.m.i.p.b.BatchInserter Copied 20000 rows to transaction table in 456.8 ms
2022-08-18T08:39:50.927-0600 INFO scheduling-3 c.h.m.i.p.b.BatchInserter Copied 20000 rows to transaction table in 257.9 ms
2022-08-18T08:39:53.420-0600 INFO scheduling-3 c.h.m.i.p.b.BatchInserter Copied 20000 rows to transaction table in 361.2 ms
2022-08-18T08:39:54.837-0600 INFO scheduling-3 c.h.m.i.p.b.BatchInserter Copied 20000 rows to transaction table in 296.5 ms
```

`transaction` table and indexes storage stats for db with data from perfnet:

```
 Table Name  |                 Index Name                 | Total Size | Total Size of all Indexes | Table Size | Index Size | Estimated table row count
-------------+--------------------------------------------+------------+---------------------------+------------+------------+---------------------------
 transaction | transaction_pkey                           | 55 GB      | 21 GB                     | 33 GB      | 3000 MB    |                 118458672
 transaction | transaction__transaction_id                | 55 GB      | 21 GB                     | 33 GB      | 5645 MB    |                 118458672
 transaction | transaction__payer_account_id_consensus_ns | 55 GB      | 21 GB                     | 33 GB      | 4212 MB    |                 118458672
 transaction | transaction__type_consensus_timestamp      | 55 GB      | 21 GB                     | 33 GB      | 4219 MB    |                 118458672
 transaction | transaction__hash                          | 55 GB      | 21 GB                     | 33 GB      | 4676 MB    |                 118458672
```

same storage stats for db with data from mainnet:

```
 Table Name  |                 Index Name                 | Total Size | Total Size of all Indexes | Table Size | Index Size | Estimated table row count
-------------+--------------------------------------------+------------+---------------------------+------------+------------+---------------------------
 transaction | transaction__type_consensus_timestamp      | 1163 GB    | 581 GB                    | 582 GB     | 139 GB     |                2413293056
 transaction | transaction__transaction_id                | 1163 GB    | 581 GB                    | 582 GB     | 138 GB     |                2413293056
 transaction | transaction__payer_account_id_consensus_ns | 1163 GB    | 581 GB                    | 582 GB     | 140 GB     |                2413293056
 transaction | transaction_pkey                           | 1163 GB    | 581 GB                    | 582 GB     | 100 GB     |                2413293056
 transaction | transaction__hash                          | 1163 GB    | 581 GB                    | 582 GB     | 64 GB      |                2413293056
```

**Checklist**

- [ ] Documented (Code comments, README, etc.)
- [x] Tested (unit, integration, etc.)